refactor the command-line-interface to use optparse
Optparse is in the standard library, and can be used in 2.4, 2.5 and 2.6.
However it will be superseded by 'argparse' module in 2.7.

The code is a bit cleaner now, and we get the help and version message for free.
However one difference is that in case no arguments are supplied the exit code
is '2', instead of '-1'.