layz cache for file content
In preperation for including content from stdin and from files, as well as easy
testing using monkey patching, a lazy cache for file content has been added.
By default get_lines_from_cache should be used, this will either return the
lines in a file if it has already been loaded, or load the file from disk.

It can also be used to initialise a file cache, using pretend names and content,
using add_lines_to_cache. This is great for monkey patching content into the
cache without actually having any files on disk. A method clear_file_cache is
also provided, just in case.

As a side node, this also means that for the use case of multiple includes, the
code should be marginally faster since the file need not be loaded from disk
twice.