new function to enable recursive inclusion
As opposed to the old include_file_recursive, this one keeps track of which
files have been included, and can thus detect loops. It works using a stack of
included files and a depth first search. If a loop is detected an Exception is
raised. It will later replace the old include_file_recursive.