refactor unit test to expose bug in file inclusion
It seems like two of the unit tests were not quite right. The unit tests in
question test to see if file inclusion is disabled inside the 'code'
environment. The problem is that they checked this works by giving input to the
'convert2beamer' function, but there is never any file inclusion in this
function. As a result the unit tests passed, even though wiki2beamer would have
failed on real[TM] input. All the file inclusion takes place earlier in the
include_file_recursive function. The unit tests were refactored to first run the
include_file_recursive function, and then the convert2beamer. They now correctly
expose the bug.